### 2023-02-20 #5119 @skaliy
Thanks a lot for the update @britta-wstnr ! We submitted a version of the manuscript back in September of 2022. Since then, we've made significant updates to the library (135 of our 167 commits are after our previous submission), expanded on the set of tutorials showcasing the library, used it in several scientific publications (Hodneland et al., 2022), and (Kaliyugarasan et al., 2023), and also as part of our teaching in regular courses and workshops. We've also had interest from other researchers (36 stars, 5 forks).

Compared to the previous version of the manuscript, the present version tries to explain a wider variety of parts of the library and also a wider variety of use cases. However, we've still kept what we believe is a significant strength of our submission: the entire manuscript is created from a Jupyter Notebook that can be run, for example, in Google Colab. In other words, entirely reproducible. We believe that fully executable papers are the future in computational fields. We hope that by providing an example, our contribution to JOSS can play a small part in helping bring forth a new default.

We believe a publication in JOSS could bring the library to the attention of more people that could benefit from its combination of simplicity and flexibility. We're great admirers of JOSS and would, if necessary, be very happy to provide additional details or make any adjustments to the manuscript you see fit to make our contribution better suit your goals.
