### 2023-03-17 #5119 @skaliy
Sorry for the late reply @britta-wstnr! I looked at the JOSS reviewers list and identified two potential reviewers who could be great candidates for evaluating our library: Fernando Andrerotti and Nina Montana-Brwon. I believe their expertise and experience would enable them to provide valuable feedback on our work.

Another excellent candidate is Fernando Pérez-García, the developer of TorchIO. Although he is not on the list, his background and skills could make him an excellent choice for reviewing our library. Here’s his website with contact info: https://www.microsoft.com/en-us/research/people/fperezgarcia/.
