### 2023-03-23 #5297 @editorialbot
[REVIEW]: fastMONAI: a low-code deep learning library for medical image analysis
**Submitting author:** @skaliy (<a href="http://orcid.org/0000-0002-0038-5540">Satheshkumar  Kaliyugarasan</a>)
**Repository:** https://github.com/MMIV-ML/fastMONAI
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
