### 2023-03-30 #5297 @NMontanaBrown
Hi @britta-wstnr and @skaliy, I have a question about the "Substantial Scholarly Effort" part of the checklist.

"In addition, JOSS requires that software should be feature-complete (i.e., no half-baked solutions), packaged appropriately according to common community standards for the programming language being used (e.g., [Python](https://packaging.python.org/), [R](https://r-pkgs.org/index.html)), and designed for maintainable extension (not one-off modifications of existing tools). **“Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable.**"

From my knowledge of MONAI, I am not convinced at the moment that this package fulfils the latter criterium. MONAI already provides all the functionality necessary that this package claims, including tutorials in the form of Jupyter notebooks, and the integration of TorchIO has already been achieved within MONAI, so this is not specifically a novelty as far as I'm aware.

@skaliy could you comment on this? and @britta-wstnr could you advise?
