### 2023-03-30 MMIV-ML/fastMONAI#34 @fepegar
Error running the notebook / paper
Hi,

I've got an error running the notebook of the paper (https://colab.research.google.com/github/MMIV-ML/fastMONAI/blob/master/paper/paper.ipynb).

```python-traceback
File ["<ipython-input-59-e08cef295d53>"](https://localhost:8080/#), line 2
    store_variables(pkl_fn='models/vars.pkl', size, reorder, resample)
                                                                     ^
SyntaxError: positional argument follows keyword argument
```

I think the line should be

```python
store_variables('models/vars.pkl', size, reorder, resample)
```

(I'm reviewing at https://github.com/openjournals/joss-reviews/issues/5297)
