### 2023-03-31 #5297 @skaliy
Dear @NMontanaBrown. Thank you for the feedback and for asking for our feedback! 👍

We are great fans of MONAI and their enormous efforts to support the entire “value chain” of AI in medical imaging, from data labeling to deployment in hospital infrastructure.

We naturally believe that there are several significant scholarly contributions in our work. Here are some of the main points:

1. fastMONAI is not only tightly integrated with MONAI and TorchIO but also fastai ([](https://www.fast.ai/)). The fastai library has a nice and flexible design, with a layered API, enabling both a low-code approach and the option to dig deep into the underlying PyTorch code. By building on fastai, we can bring these features into the medical imaging domain.

2. In a sense, we provide a wrapper. Still, it is a wrapper of three very powerful libraries, MONAI, fastai, and TorchIO (and they, of course, are, in a sense, wrappers of other libraries, like PyTorch, SimpleITK, NiBabel, PyDicom, and so on). But we’ve put significant effort into constructing a practical, robust wrapper, which is also well-documented (with the documentation automatically synchronized with code updates).

3. In general, even if the functionality and objectives of Project MONAI, nnU-Net, DeepMedic, and others overlap with what we’re trying to do in fastMONAI, we believe that the community benefits from a plethora of approaches.

4. We have found fastMONAI very useful for our research, enabling quick prototyping while having lower-level access and for teaching in university courses and workshops.

5. We think that the development process of fastMONAI (and fastai) using nbdev, and the fact that our paper manuscript is executable as a Jupyter Notebook, could be of interest to the broader open-source community served by JOSS.

6. Finally, I’m handing in my thesis in a week, so the past two months have been very hectic. But once some time is freed up, we have multiple plans for the further development of fastMONAI and adding new tutorials. We will also continue our work on integrating fastMONAI with our hospital’s (open source) research PACS system (ref. [](https://bora.uib.no/bora-xmlui/handle/11250/3021987)).

Again, we absolutely see where you are coming from, given how great the MONAI framework is, and we greatly appreciate that you allowed us to argue for our work.🙂
