### 2023-03-31 394620e
fix minor error in the notebook/paper
