### 2023-03-31 MMIV-ML/fastMONAI#34 @skaliy
Hi,

Thanks for letting us know. It's been fixed now.
