### 2023-03-31 #5297 @britta-wstnr
Thank you for the feedback @NMontanaBrown and thank you for your response @skaliy!

It would be very useful if @fepegar and @patrickraoulphilipp could give their opinion/assessment of this as well (and @NMontanaBrown of course feel free to further chime in or comment on @skaliy's reply).

Based on @skaliy's reply, I would say the question to answer is whether this statement from our documentation is applicable (see added emphasis):
"Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or **makes addressing research challenges significantly better (e.g., faster, easier, simpler).**"
(from https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
