### 2023-04-04 #5297 @fepegar
# Review

Hi, all. I've performed a thorough pass on the paper and repository. I got an error while running the notebook/paper, which I reported, as you might have seen above.

I've been waiting for your replies to @NMontanaBrown's comment, with which I strongly agree.

## Summary

**I recommend that the paper is rejected**.

From my point of view, fastMONAI comprises a small set of convenience tools and thin wrappers that have been useful for some of the author's research projects, but I do not think it addresses research challenges for a general audience of researchers in the field.

## Replies to author

I will reply to some of @skaliy's comments below.

### Wrapper

> 2. In a sense, we provide a wrapper. Still, it is a wrapper of three very powerful libraries, MONAI, fastai, and TorchIO (and they, of course, are, in a sense, wrappers of other libraries, like PyTorch, SimpleITK, NiBabel, PyDicom, and so on). But we’ve put significant effort into constructing a practical, robust wrapper, which is also well-documented (with the documentation automatically synchronized with code updates).

I think that most of the features of e.g. MONAI and TorchIO _use_ or _are built on top of_ PyTorch, SimpleITK, etc., as opposed to being wrappers of those lower-level libraries.

Here are some samples from the documentation:

Here is the code for the second screenshot:

```python
# %% ../nbs/03_vision_augment.ipynb 8
def do_pad_or_crop(o, target_shape, padding_mode, mask_name, dtype=torch.Tensor):

    pad_or_crop = tio.CropOrPad(target_shape=target_shape, padding_mode=padding_mode, mask_name=mask_name)
    return dtype(pad_or_crop(o))

# %% ../nbs/03_vision_augment.ipynb 9
class PadOrCrop(DisplayedTransform):
    '''Resize image using TorchIO `CropOrPad`.'''

    order=0
    def __init__(self, size, padding_mode=0, mask_name=None):
        if not is_listy(size): size=[size,size,size]
        self.size, self.padding_mode, self.mask_name = size, padding_mode, mask_name

    def encodes(self, o:(MedImage, MedMask)):
        return do_pad_or_crop(o,target_shape=self.size, padding_mode=self.padding_mode, mask_name=self.mask_name, dtype=type(o))
```

I would argue that many of these transforms are _very thin_ wrappers, with hard-coded arguments that might not generalize to many applications. Moreover, they are not "well-documented" as claimed.

### Adoption by the community

> In general, even if the functionality and objectives of Project MONAI, nnU-Net, DeepMedic, and others overlap with what we’re trying to do in fastMONAI, we believe that the community benefits from a plethora of approaches.

> We have found fastMONAI very useful for our research, enabling quick prototyping while having lower-level access and for teaching in university courses and workshops.

It's not clear whether the community is benefiting from fastMONAI.

[![Star History Chart](https://api.star-history.com/svg?repos=MMIV-ML/fastMONAI&type=Date)](https://star-history.com/#MMIV-ML/fastMONAI&Date)

It seems that most of the stars on the GitHub repo (27/38) were obtained within the first month of its existence (September 2022), and the count has been growing very slowly since then, suggesting little adoption from the community.

[There are no dependents of the library on GitHub](https://github.com/MMIV-ML/fastMONAI/network/dependents).

Five papers are cited in the "Research projects using fastMONAI" section of the paper.
I believe that 1) @skaliy is an author in all of them, and 2) fastMONAI is not mentioned in the papers.

### nbdev

> We think that the development process of fastMONAI (and fastai) using nbdev, and the fact that our paper manuscript is executable as a Jupyter Notebook, could be of interest to the broader open-source community served by JOSS.

nbdev is not a standard framework in widely used libraries in the field.
Writing code in notebooks and generating the Python library programmatically from them makes the repository harder to navigate and understand, from my point of view.
Moreover, there is not a clear or standard structure for unit tests.

I am not sure that the fact that the paper is executable as a Jupyter notebook is of interest to the broader open-source community.

I suppose many of these issues are inherited from https://github.com/fastai/nbdev-template, of which fastMONAI is a fork.

## Other considerations

### Name

The name `fastMONAI` (fastai + MONAI?) suggests that the library is a faster version of MONAI, which is not the case. I suggest using a different name.

### Coding style

Consider the following snippet from the library.

```python
# %% ../nbs/03_vision_augment.ipynb 37
class RandomMotion(DisplayedTransform):
    '''Apply TorchIO `RandomMotion`.'''

    split_idx,order=0,1

    def __init__(self, degrees=10, translation=10, num_transforms=2, image_interpolation='linear', p=0.5):
        self.degrees,self.translation, self.num_transforms, self.image_interpolation, self.p = degrees,translation, num_transforms, image_interpolation, p

    def encodes(self, o:(MedImage)): return MedImage.create(_do_rand_motion(o, degrees=self.degrees,translation=self.translation, num_transforms=self.num_transforms, image_interpolation=self.image_interpolation, p=self.p))
    def encodes(self, o:(MedMask)):return o
```

It would greatly help to follow a common coding style, such as [PEP8](https://www.python.org/dev/peps/pep-0008/), to improve readability and maintainability (and reviewing). Some better practices would be:

- Use more descriptive variable names
- Use shorter lines
- Avoid parentheses for typing annotations
- Add spaces after commas and around equal signs

I would recommend a combination of `pre-commit`, `black` and `flake8` to enforce a consistent, standard coding style.

Certain typing annotations are invalid, as in the following snippet.

```python
class PadOrCrop(DisplayedTransform):
    '''Resize image using TorchIO `CropOrPad`.'''

    order=0
    def __init__(self, size, padding_mode=0, mask_name=None):
        if not is_listy(size): size=[size,size,size]
        self.size, self.padding_mode, self.mask_name = size, padding_mode, mask_name

    def encodes(self, o:(MedImage, MedMask)):
        return do_pad_or_crop(o,target_shape=self.size, padding_mode=self.padding_mode, mask_name=self.mask_name, dtype=type(o))
```

The following line, very typical in `fastai`, I believe, makes reviewing also quite difficult:

```python
from fastai.data.all import *
```

It is also used in the paper/notebook/tutorial, making it difficult to understand where some functions are classes come from.

Here is one of the many resources explaining why this is typically considered bad practice: https://www.geeksforgeeks.org/why-import-star-in-python-is-a-bad-idea/

### Paper

The paper does not clearly justify the need for this library.

Most of the content is a tutorial, which would belong in the documentation instead.

The "IXI Tiny" dataset is a TorchIO class, which is not referenced in the paper.

From the paper:

> "To ease further extensions of our library through contributions, we have added a short guide on how to contribute to the project"

However, the `CONTRIBUTING.md` file only contains the following text:

> # How to contribute
> fastMONAI follows the same contribution policy as fastai: https://github.com/fastai/nbdev/blob/master/CONTRIBUTING.md
