+++ paper/paper.bib
@@ -80,7 +80,9 @@
 }
 
 @misc{ixi,
+  author = {Alansary, Amir and others},
   title = {IXI Dataset},
+  publisher = {Biomedical Image Analysis Group, Imperial College London},
   howpublished = {\url{http://brain-development.org/ixi-dataset/}}
 }
 
