+++ paper/paper.md
@@ -7,23 +7,20 @@ tags:
   - fastai
   - MONAI
   - torchIO
-authors:
-  - name: Satheshkumar Kaliyugarasan
-    orcid: 0000-0002-0038-5540
-    affiliation: "1,2" 
-    
-  - name: Alexander Selvikvåg Lundervold
-    orcid: 0000-0001-8663-4247
-    affiliation: "1,2" 
-affiliations:
- - name: Department of Computer Science, Electrical Engineering and Mathematical Sciences, Western Norway University of Applied Sciences, Bergen, Norway
-   index: 1
- - name: Mohn Medical Imaging and Visualization Centre, Department of Radiology, Haukeland University Hospital, Bergen, Norway
-   index: 2
-date: February 2023
+
+author:
+- Satheshkumar Kaliyugarasan
+- Alexander Selvikvåg Lundervold
+  
+  \vspace{0.5cm}
+- Dept. of Computer Science, Electrical Engineering and Mathematical Sciences, HVL
+- MMIV, Dept. of Radiology, Haukeland University Hospital
+date: April 2023
+geometry: margin=2cm
+colorlinks: true
 bibliography: paper.bib
 ---
-[![Google Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/MMIV-ML/fastMONAI/blob/master/paper/paper.ipynb)
+[![Google Colab](paper_files/colab-badge.svg){ width=20% }](https://colab.research.google.com/github/MMIV-ML/fastMONAI/blob/master/paper/paper.ipynb)
 
 # Summary
 
@@ -64,7 +61,7 @@ from fastMONAI.vision_all import *
 To demonstrate the use of fastMONAI. we download the NoduleMNIST3D dataset from MedMNIST v2 [@yang2023medmnist], a dataset containing lung nodules with labels indicating whether the nodules are benign (b) or malignant (m):
 
 ```python
-df, _ = download_NoduleMNIST3D(max_workers=8)
+df, _ = download_NoduleMNIST3D(max_workers = 8)
 ```
 
 ### Inspecting the data
@@ -82,7 +79,7 @@ print(df.head(1).to_markdown())
 In fastMONAI, various data augmentation techniques are available for training vision models, and they can also optionally be applied during inference. The following code cell specifies a list of transformations to be applied to the items in the training set. The complete list of available transformations in the library can be found at [https://fastmonai.no/vision_augment](https://fastmonai.no/vision_augment).
 
 ```python
-item_tfms=[PadOrCrop(size=28), RandomAffine(degrees=35, isotropic=True), 
+item_tfms = [PadOrCrop(size = 28), RandomAffine(degrees = 35, isotropic = True), 
            ZNormalization()]
 ```
 
@@ -92,14 +89,14 @@ In the following line, we call the ` ImageDataLoaders.from_df` factory method, w
 Here, we pass the processed DataFrame, define the columns for the images `fn_col` and the labels `label_col`, some transforms `item_tfms`, voxel spacing `resample`, and the batch size `bs`.
 
 ```python
-dls = MedImageDataLoaders.from_df(df, fn_col='img_path', label_col='labels', 
-                                  item_tfms=item_tfms, resample=1, bs=64)
+dls = MedImageDataLoaders.from_df(df, fn_col = 'img_path', label_col = 'labels', 
+                                  item_tfms = item_tfms, resample = 1, bs = 64)
 ```
 
 We can now take a look at a batch of images in the training set using `show_batch` :
 
 ```python
-dls.show_batch(max_n=2, anatomical_plane=2)
+dls.show_batch(max_n = 2, anatomical_plane = 2)
 ```
 
 ![](paper_files/output_22_0.png){ width=35% }
[…]
