+++ paper/README.md
@@ -1,3 +1,3 @@
 | PDF | Notebook                                                                                                                                                                            | Colab                                                                                                                                                                 |
 |----:|-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-|   - | [![nbviewer](https://raw.githubusercontent.com/jupyter/design/master/logos/Badges/nbviewer_badge.svg)](https://nbviewer.org/github/MMIV-ML/fastMONAI/blob/master/paper/paper.ipynb) | [![Google Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/MMIV-ML/fastMONAI/blob/master/paper/paper.ipynb) |
+| [![pdf](https://github.com/FortAwesome/Font-Awesome/blob/6.x/svgs/regular/file-pdf.svg)](https://github.com/MMIV-ML/fastMONAI/blob/master/paper/paper.pdf) | [![nbviewer](https://raw.githubusercontent.com/jupyter/design/master/logos/Badges/nbviewer_badge.svg)](https://nbviewer.org/github/MMIV-ML/fastMONAI/blob/master/paper/paper.ipynb) | [![Google Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/MMIV-ML/fastMONAI/blob/master/paper/paper.ipynb) |
