### 2023-04-16 #5297 @patrickraoulphilipp
Hi all,

first of all thank you for you efforts on the fastMONAI library, it is very interesting and promising! Here are some initial thoughts and questions of mine after reading the manuscript and resources, as well as testing the framework.

I agree with prior mentioned points, especially:

(1) No clear distinction / explanation of added value of the project wrt MONAI and TorchIO. It would highly beneficial to add clear cases where fastMONAI eases prototyping and developing. It would be help the raised question on deciding if sufficient added value is provided for the journal. Which additional model architectures / trained models can I use now? How much glue code would be necessary otherwise? Where does fastMONAI go beyond the existing possibility to use MONAI and TorchIO together?

(2) Use of * imports as used in fastai / inconsistent code formatting. As said before, this does make understanding and debugging code difficult. But I am glad to hear that upcoming refactorings might get rid of this.

I started testing the software on collab and locally (using Windows). I had no problems to go though the example notebooks via collab, but ran into problems locally in Windows (when installing from source and when training a model). Maybe I did oversee, but which Python versions are supported / recommended? I will try again.
