### 2023-05-08 #5297 @skaliy
Thank you for your valuable feedback @patrickraoulphilipp!

To address your concerns:

1. Added value of fastMONAI compared to MONAI and TorchIO: We acknowledge that a clearer distinction and explanation of the added value of fastMONAI compared to MONAI and TorchIO is needed. We plan to include more detailed use cases and examples where fastMONAI eases prototyping and development.
2. At the moment, fastchan is not supported on Windows. To resolve this, replace -c fastchan in the installation process with -c pytorch -c nvidia -c fastai. We will work on fixing this problem and provide clearer guidance on Python versions supported and recommended for our library.

@britta-wstnr, we regret to inform you that, given the points raised and the bar set by @fepegar, and @NMontanaBrown, we don't believe it will be possible to sufficiently improve the project's overall quality within the review timeline.

Once again, thank you all for your valuable feedback. We will continue working on improving the fastMONAI library and address your concerns.
